requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
channels==4.2.2
daphne==4.1.2
pytz==2025.2
